Rename historyOperations components to multi-word, and for clarity matching imports and usage
