+++ client/src/components/History/CurrentHistory/HistoryOperations/DefaultOperations.vue
