+++ client/src/components/History/CurrentHistory/HistoryPanel.vue
@@ -148,7 +148,7 @@ import ExpandedItems from "components/History/Content/ExpandedItems";
 import SelectedItems from "components/History/Content/SelectedItems";
 import ListingLayout from "components/History/Layout/ListingLayout";
 import HistoryCounter from "./HistoryCounter";
-import HistoryOperations from "./HistoryOperations/Index";
+import HistoryOperations from "./HistoryOperations/HistoryOperations";
 import HistoryDetails from "./HistoryDetails";
 import HistoryDropZone from "./HistoryDropZone";
 import HistoryEmpty from "./HistoryEmpty";
